# internal/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nops API",
    "version": "1",
    "license": { "name": "Apache 2.0", "identifier": "Apache-2.0" },
    "description": "Read jobs and deployments, act on them, and manage ACL policies, binding rules and tokens. A token does what its ACL policies allow, with the dashboard's rules."
  },
  "servers": [{ "url": ".." }],
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/api/acl/tokens": {
      "get": {
        "summary": "Every token",
        "operationId": "listACLTokens",
        "responses": {
          "200": {
            "description": "The tokens, newest first.",
            "content": { "application/json": { "schema": { "type": "object", "required": ["tokens"], "properties": { "tokens": { "type": "array", "items": { "$ref": "#/components/schemas/ACLToken" } } } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "summary": "Create a token",
        "operationId": "createACLToken",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLTokenBody" } } }
        },
        "responses": {
          "201": {
            "description": "The token, with its secret.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewACLToken" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/acl/tokens/{accessor_id}": {
      "get": {
        "summary": "One token",
        "operationId": "getACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "200": {
            "description": "The token.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLToken" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Revoke a token",
        "operationId": "revokeACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "204": { "description": "Revoked." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "A token from the dashboard's Administration page." }
    },
    "parameters": {
...
      "AccessorID": { "name": "accessor_id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is not what the endpoint takes.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "No valid token: it is missing, unknown, expired, or revoked.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The token is valid, and its ACL policies do not allow this.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The job, deployment, ACL policy or token does not exist.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ServerError": {
        "description": "Nops hit an error and logged it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "ACLToken": {
        "type": "object",
        "required": ["accessor_id", "name", "type", "origin", "policies", "created_at"],
        "properties": {
          "accessor_id": { "type": "string", "description": "Public: it names the token in logs and the audit, and grants nothing." },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["management", "client"] },
          "origin": { "type": "string", "enum": ["login", "created"], "description": "A login made a session token; a person or a token created the others." },
          "identity": { "type": "string", "description": "The person a session token stands for. Absent for a created token." },
          "policies": { "type": "array", "items": { "type": "string" }, "description": "The ACL policies a client token carries, by name." },
          "created_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time", "description": "Absent for a token that never expires." },
          "creator_accessor_id": { "type": "string", "description": "The token that made this one. It stays after that token is gone." },
          "creator_name": { "type": "string" },
          "creator_identity": { "type": "string", "description": "The person who created the token, when it was one." }
        }
      },
      "NewACLToken": {
        "allOf": [
          { "$ref": "#/components/schemas/ACLToken" },
          {
            "type": "object",
            "required": ["secret"],
            "properties": { "secret": { "type": "string", "description": "The secret. Nops shows it once and keeps only its hash." } }
          }
        ]
      },
      "ACLTokenBody": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "type": { "type": "string", "enum": ["management", "client"], "description": "Default client." },
          "policies": { "type": "array", "items": { "type": "string" }, "description": "The ACL policies of a client token. Each must exist." },
          "expires_in": { "type": "string", "description": "A duration such as 720h. Empty: the token never expires." }
        }
      },
...
    }
  }
}